refactor: split ai research runtime
